add test for task code replace
